requirements.txt
selenium>=4.0.0
python-dotenv>=0.21.0
webdriver-manager>=3.8.0
requests>=2.28.0